extract: toggleSelectedRepository in redhatRepositories
toggleSelectedRepository has been extracted into its
own function and composed in a hook.
