+++ src/features/createAndEditTemplate/redhatRepositories/core/ports.ts
@@ -1,0 +1,5 @@
+import { AdditionalUUID } from 'features/createAndEditTemplate/shared/types/types';
+
+// ports input
+export type ToggleSelectedAdditionalRepository = (uuid: AdditionalUUID) => void;
+// display redhat repositories
