+++ src/features/createAndEditTemplate/redhatRepositories/core/use-cases/toggleSelectedRepository.ts
@@ -1,0 +1,20 @@
+import { useCallback } from 'react';
+import { useTemplateRequestApi } from 'features/createAndEditTemplate/workflow/store/TemplateStore';
+import { ToggleSelectedAdditionalRepository } from '../ports';
+
+export const useToggleSelectedRepository = () => {
+  const { setAdditionalUUIDs } = useTemplateRequestApi();
+
+  const toggleSelected: ToggleSelectedAdditionalRepository = useCallback((clickedUuid) => {
+    setAdditionalUUIDs((previous) => {
+      const isInPrevious = previous.includes(clickedUuid);
+      if (isInPrevious) {
+        return previous.filter((uuid) => uuid !== clickedUuid);
+      } else {
+        return [...previous, clickedUuid];
+      }
+    });
+  }, []);
+
+  return toggleSelected;
+};
